feat: Be as good as possible while being as least bad as possible
