+++ examples/next-fastapi/app/(examples)/01-chat-text/page.tsx
@@ -1,12 +1,12 @@
 'use client';
 
 import { Card } from '@/app/components';
-import { useChat, ReactChatStore } from '@ai-sdk/react';
+import { useChat, ChatStore } from '@ai-sdk/react';
 import { TextStreamChatTransport } from 'ai';
 
 export default function Page() {
   const { messages, input, handleSubmit, handleInputChange, status } = useChat({
-    chatStore: new ReactChatStore({
+    chatStore: new ChatStore({
       transport: new TextStreamChatTransport({
         api: '/api/chat?protocol=text',
       }),
