+++ examples/sveltekit-openai/src/routes/chat/[id]/+page.svelte
@@ -3,14 +3,15 @@
   import ArrowUp from '$lib/components/icons/arrow-up.svelte';
   import Button from '$lib/components/ui/button/button.svelte';
   import { Textarea } from '$lib/components/ui/textarea/index.js';
-  import { Chat, defaultChatStore } from '@ai-sdk/svelte';
+  import { Chat } from '@ai-sdk/svelte';
+  import { DefaultChatTransport } from 'ai';
 
-  const chat = new Chat(() => ({
-    chatId: page.params.id,
-    chatStore: defaultChatStore({
+  const chat = new Chat({
+    id: page.params.id,
+    transport: new DefaultChatTransport({
       api: '/api/chat',
-      maxSteps: 5,
     }),
+    maxSteps: 5,
     // run client-side tools that are automatically executed:
     async onToolCall({ toolCall }) {
       // artificial 2 second delay
@@ -21,7 +22,7 @@
         return cities[Math.floor(Math.random() * cities.length)];
       }
     },
-  }));
+  });
 
   const disabled = $derived(chat.status !== 'ready');
