+++ packages/ai/src/ui/get-tool-invocations.ts
@@ -1,9 +1,13 @@
 import { ToolInvocation, ToolInvocationUIPart, UIMessage } from './ui-messages';
 
-export function getToolInvocations(message: UIMessage): ToolInvocation[] {
-  return message.parts
-    .filter(
-      (part): part is ToolInvocationUIPart => part.type === 'tool-invocation',
-    )
-    .map(part => part.toolInvocation);
+export function getToolInvocations(
+  message: UIMessage | undefined,
+): ToolInvocation[] {
+  return (
+    message?.parts
+      .filter(
+        (part): part is ToolInvocationUIPart => part.type === 'tool-invocation',
+      )
+      .map(part => part.toolInvocation) ?? []
+  );
 }
