+++ packages/react/src/index.ts
@@ -1,4 +1,4 @@
 export * from './use-chat';
 export * from './use-completion';
 export * from './use-object';
-export { defaultChatStore, ReactChatStore } from './chat-store';
+export { defaultChatStore, ChatStore } from './chat-store';
