+++ packages/svelte/src/context-provider.ts
@@ -1,8 +1,3 @@
-import { ChatStore, defaultChatStore } from 'ai';
-import {
-  setChatStoreContext,
-  SvelteStateManager,
-} from './chat-store.svelte.js';
 import {
   KeyedCompletionStore,
   setCompletionContext,
@@ -12,22 +7,10 @@ import {
   setStructuredObjectContext,
 } from './structured-object-context.svelte.js';
 
-export function createAIContext(chatStore?: ChatStore) {
-  createChatStoreContext(chatStore);
-
+export function createAIContext() {
   const completionStore = new KeyedCompletionStore();
   setCompletionContext(completionStore);
 
   const objectStore = new KeyedStructuredObjectStore();
   setStructuredObjectContext(objectStore);
-}
-
-export function createChatStoreContext(chatStore?: ChatStore) {
-  setChatStoreContext(
-    chatStore ??
-      defaultChatStore({
-        api: '/api/chat',
-        StateManager: SvelteStateManager,
-      }),
-  );
 }
